@@ -19,6 +19,7 @@
     "base.js",
     "typescript.js"
   ],
+  "type": "module",
   "author": "Nathaniel Tucker",
   "license": "BSD-3-Clause",
   "homepage": "https://github.com/ntucker/anansi/tree/master/packages/eslint-plugin#README",
@@ -28,31 +29,39 @@
     "directory": "packages/eslint-plugin"
   },
   "engines": {
-    "node": "^16.0.0 || >=18.0.0"
+    "node": "^18.18.0 || >=20.0.0"
   },
   "dependencies": {
     "@babel/eslint-parser": "^7.25.1",
-    "@typescript-eslint/parser": "^8.4.0",
+    "@babel/eslint-plugin": "^7.25.1",
+    "@eslint/js": "^9.9.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.6.3",
     "eslint-import-resolver-webpack": "^0.13.9",
+    "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.10.0",
-    "semver": "^7.6.3"
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "globals": "^15.9.0",
+    "semver": "^7.6.3",
+    "typescript-eslint": "^8.4.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.25.2",
+    "@types/babel__core": "^7",
+    "@types/eslint": "^9",
     "@types/node": "^22.0.0",
-    "@types/semver": "^7"
+    "@types/semver": "^7",
+    "eslint": "^9.9.1",
+    "prettier": "^3.3.3",
+    "typescript": "^5.5.4",
+    "webpack": "^5.94.0"
   },
   "peerDependencies": {
     "@babel/core": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^7.0.0 || ^8.0.0",
-    "eslint": "^8.56.0 || ^9.0.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^4.0.0 || ^5.0.0",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "^2.2.1 || ^3.0.0",
+    "eslint": "^9.0.0",
+    "prettier": "^3.0.0",
     "typescript": "^4.0.0 || ^5.0.0",
     "webpack": "^4.0.0 || ^5.0.0"
   },
