@@ -1,6 +1,9 @@
-module.exports = {
+import javascript from './base.js';
+import typescript from './typescript.js';
+
+export default {
   configs: {
-    typescript: require('./typescript'),
-    javascript: require('./base'),
+    typescript,
+    javascript,
   },
 };
